refactor: unify exception package
