+++ src/greeks/finite_difference.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import numpy as np
 from numpy.typing import NDArray
 
-from ..utils.exceptions import ValidationError
+from src.exceptions import ValidationError
 from .base import GreeksCalculator
