+++ src/utils/process_validators.py
@@ -8,7 +8,7 @@ import numpy as np
 from numpy.typing import NDArray
 
 from ..validation import validate_positive, validate_non_negative, validate_probability
-from .exceptions import ValidationError
+from src.exceptions import ValidationError
 
 
 def validate_feller_condition(
